pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "karna-vlm"
version = "0.1.0"
description = "Karna VLM — A compact, customizable vision-language model platform"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Karna AI", email = "team@karna.ai"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "tokenizers>=0.15.0",
    "safetensors>=0.4.0",
    "Pillow>=10.0.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.25.0",
    "datasets>=2.16.0",
    "einops>=0.7.0",
    "timm>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
train = [
    "deepspeed>=0.12.0",
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
]
eval = [
    "pycocoevalcap>=1.2",
    "nltk>=3.8.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
]
serve = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]
export = [
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "optimum>=1.14.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
all = ["karna-vlm[train,eval,serve,export,dev]"]

[project.scripts]
karna = "karna_vlm.cli:app"

[project.urls]
Homepage = "https://github.com/Viraj0518/Karna-OCR"
Documentation = "https://github.com/Viraj0518/Karna-OCR/tree/main/docs"
Repository = "https://github.com/Viraj0518/Karna-OCR"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
